Force dependencies to use parse5@4.0.0 (earlier versions are broken)
parse5 prior to v4.0.0 included @types/node in its runtime dependencies,
which creates conflicts with react-native when compiling or linting
TypeScript.

parse5 is currently used by jest dependencies (jsdom) and enzyme
dependencies (cheerio). cheerio requires 3.x.x, and jsdom specifically
requires 4.0.0. However, cheerio appears to work with 4.0.0 as well, as
the only changes between 3.x.x and 4.0.0 appear to be to node-specific
APIs. Therefore forcing it to use 4.0.0 as well appears to fix the
problem so far.