Refactor common async action handling pattern into helper class
`Action` allows events in the lifecycle of asynchronous commands to be
represented in a consistent manner across the codebase, and provides a
monadic-esque API for easily mapping the success/failure data. It also
allows the mapping functions to add updates to the reducer state, which
streamlines the conceptual clutter of reducers, especially when combined
with the `asyncReducer` helper. Finally, it reduces the number of action
types needed, by representing all the async action types itself, rather
than needing them to be explicitly defined.